+++ paq8pxd.cpp
@@ -144,6 +144,7 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "filters/cdfilter.hpp" 
 #include "filters/giffilter.hpp" 
 #include "filters/szddfilter.hpp" 
+#include "filters/base64filter.hpp" 
 
 Img24Filter img24("image 24bit");
 Img32Filter img32("image 32bit");
@@ -155,6 +156,7 @@ MDFFilter    mdff("mdf");
 CDFilter    cdff("cd");
 gifFilter    giff("gif");
 szddFilter    szddf("szdd");
+base64Filter  base64f("base64");
 
 Streams streams;
 /////////////////////// Global context /////////////////////////
@@ -290,10 +292,6 @@ deth=(header_len),detd=(data_len),info=(-1),info2=(-1),\
 #define NES_DET(type,start_pos,header_len,base64len) return dett=(type),\
 deth=(-1),detd=int(base64len),\
  in->setpos(start+start_pos),DEFAULT
-
-inline bool is_base64(unsigned char c) {
-    return (isalnum(c) || (c=='+') || (c=='/')|| (c==10) || (c==13));
-}
 
 inline bool is_base85(unsigned char c) {
     return (isalnum(c) || (c==13) || (c==10) || (c=='y') || (c=='z') || (c>='!' && c<='u'));
@@ -4337,143 +4335,6 @@ int encode_txt_wit(File* in, File* out, U64 len) {
 }
 // end WIT
 
-// decode/encode base64 
-static const char  table1[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-bool isbase64(unsigned char c) {
-    return (isalnum(c) || (c == '+') || (c == '/')|| (c == 10) || (c == 13));
-}
-
-int decode_base64(File*in, int size, File*out, FMode mode, U64 &diffFound){
-    U8 inn[3];
-    int i, len=0, blocksout = 0;
-    int fle=0;
-    int linesize=0; 
-    int outlen=0;
-    int tlf=0,g=0;
-    linesize=in->getc();
-    outlen=in->getc();
-    outlen+=(in->getc()<<8);
-    outlen+=(in->getc()<<16);
-    tlf=(in->getc());
-    outlen+=((tlf&63)<<24);
-    Array<U8,1> ptr((outlen>>2)*4+10);
-    tlf=(tlf&192);
-    if (tlf==128)       tlf=10;        // LF: 10
-    else if (tlf==64)   tlf=13;        // LF: 13
-    else                tlf=0;
- 
-    while(fle<outlen){
-        len=0;
-        for(i=0;i<3;i++){
-            int c=in->getc();
-            if(c!=EOF) {
-                inn[i]=c;
-                len++;
-            }
-            else {
-                inn[i] = 0,g=1;
-            }
-        }
-        if(len){
-            U8 in0,in1,in2;
-            in0=inn[0],in1=inn[1],in2=inn[2];
-            ptr[fle++]=(table1[in0>>2]);
-            ptr[fle++]=(table1[((in0&0x03)<<4)|((in1&0xf0)>>4)]);
-            ptr[fle++]=((len>1?table1[((in1&0x0f)<<2)|((in2&0xc0)>>6)]:'='));
-            ptr[fle++]=((len>2?table1[in2&0x3f]:'='));
-            blocksout++;
-        }
-        if(blocksout>=(linesize/4) && linesize!=0){ //no lf if linesize==0
-            if( blocksout &&  !in->eof() && fle<=outlen) { //no lf if eof
-                if (tlf) ptr[fle++]=(tlf);
-                else ptr[fle++]=13,ptr[fle++]=10;
-            }
-            blocksout = 0;
-        }
-        if (g) break; //if past eof, break
-    }
-    //Write out or compare
-    if (mode==FDECOMPRESS){
-            out->blockwrite(&ptr[0],   outlen  );
-        }
-    else if (mode==FCOMPARE){
-    for(i=0;i<outlen;i++){
-        U8 b=ptr[i];
-        U8 c=out->getc();
-            if (b!=c && !diffFound) diffFound= out->curpos();
-        }
-    }
-    return outlen;
-}
-   
-inline char valueb(char c){
-       const char *p = strchr(table1, c);
-       if(p) {
-          return p-table1;
-       } else {
-          return 0;
-       }
-}
-
-void encode_base64(File* in, File* out, int len) {
-  int in_len = 0;
-  int i = 0;
-  int j = 0;
-  int b=0;
-  int lfp=0;
-  int tlf=0;
-  char src[4];
-  int b64mem=(len>>2)*3+10;
-  Array<U8,1> ptr(b64mem);
-  int olen=5;
-
-  while (b=in->getc(),in_len++ , ( b != '=') && is_base64(b) && in_len<=len) {
-    if (b==13 || b==10) {
-       if (lfp==0) lfp=in_len ,tlf=b;
-       if (tlf!=b) tlf=0;
-       continue;
-    }
-    src[i++] = b; 
-    if (i ==4){
-          for (j = 0; j <4; j++) src[j] = valueb(src[j]);
-          src[0] = (src[0] << 2) + ((src[1] & 0x30) >> 4);
-          src[1] = ((src[1] & 0xf) << 4) + ((src[2] & 0x3c) >> 2);
-          src[2] = ((src[2] & 0x3) << 6) + src[3];
-    
-          ptr[olen++]=src[0];
-          ptr[olen++]=src[1];
-          ptr[olen++]=src[2];
-      i = 0;
-    }
-  }
-
-  if (i){
-    for (j=i;j<4;j++)
-      src[j] = 0;
-
-    for (j=0;j<4;j++)
-      src[j] = valueb(src[j]);
-
-    src[0] = (src[0] << 2) + ((src[1] & 0x30) >> 4);
-    src[1] = ((src[1] & 0xf) << 4) + ((src[2] & 0x3c) >> 2);
-    src[2] = ((src[2] & 0x3) << 6) + src[3];
-
-    for (j=0;(j<i-1);j++) {
-        ptr[olen++]=src[j];
-    }
-  }
-  ptr[0]=lfp&255; //nl lenght
-  ptr[1]=len&255;
-  ptr[2]=len>>8&255;
-  ptr[3]=len>>16&255;
-  if (tlf!=0) {
-    if (tlf==10) ptr[4]=128;
-    else ptr[4]=64;
-  }
-  else
-      ptr[4]=len>>24&63; //1100 0000
-  out->blockwrite(&ptr[0],   olen  );
-}
 
 //base85
 int powers[5] = {85*85*85*85, 85*85*85, 85*85, 85, 1};
@@ -4672,7 +4533,7 @@ void transform_encode_block(Filetype type, File*in, U64 len, Encoder &en, int in
             eolf.encode(in, tmp, len,info&1);
             diffFound=eolf.diffFound;
         }
-        else if (type==BASE64) encode_base64(in, tmp, int(len));
+        else if (type==BASE64) base64f.encode(in, tmp, len,0);
         else if (type==UUENC) encode_uud(in, tmp, int(len),info);
         else if (type==BASE85) encode_ascii85(in, tmp, int(len));
         else if (type==SZDD) {
@@ -4700,7 +4561,9 @@ void transform_encode_block(Filetype type, File*in, U64 len, Encoder &en, int in
         
         if (type==BZIP2 || type==ZLIB || type==GIF || type==MRBR|| type==MRBR4|| type==RLE|| type==LZW||type==BASE85 ||type==BASE64 || type==UUENC|| type==DECA|| type==ARM || (type==WIT||type==TEXT || type==TXTUTF8 ||type==TEXT0)||type==EOLTEXT ){
          in->setpos(begin);
-        if (type==BASE64 ) decode_base64(tmp, int(tmpsize), in, FCOMPARE, diffFound);
+        if (type==BASE64 ) {
+            diffFound=base64f.CompareFiles(tmp,in, tmpsize, uint64_t(info), FCOMPARE);
+        }
         else if (type==UUENC ) decode_uud(tmp, int(tmpsize), in, FCOMPARE, diffFound);
         else if (type==BASE85 ) decode_ascii85(tmp, int(tmpsize), in, FCOMPARE, diffFound);
         else if (type==ZLIB && !diffFound) decode_zlib(tmp, int(tmpsize), in, FCOMPARE, diffFound);
@@ -5136,7 +4999,10 @@ U64 decompressStreamRecursive(File*out, U64 size, Encoder& en, FMode mode, int i
             decompressStreamRecursive(&tmp, len, en, FDECOMPRESS, it+1, s1+i, s2-len);
             if (mode!=FDISCARD) {
                 tmp.setpos(0);
-                if (type==BASE64) len=decode_base64(&tmp, int(len), out, mode, diffFound);
+                if (type==BASE64) {
+                    diffFound=base64f.CompareFiles(&tmp,out,len,uint64_t(info),mode);
+                    len=base64f.fsize; // get decoded size
+                }
                 else if (type==UUENC)  len=decode_uud(&tmp, int(len), out, mode, diffFound);
                 else if (type==BASE85) len=decode_ascii85(&tmp, int(len), out, mode, diffFound);
                 else if (type==SZDD)  {
